Added some new definitions in opt.
Also added a harness program, it just adds a bunch of tests
from the command line and the user can compile them out by defining
the units that are wanted.

Likewise, if you want to run all the tests, don't define anything
and run --all when you run the harness ( single-test is the name of
the binary ).